requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Flask-SQLAlchemy==2.4.4
mysql-connector==2.2.9
requests==2.25.1